pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "temper-ai"
version = "1.0.0"
description = "Temper AI agent framework with self-improvement and full observability"
license = "Apache-2.0"
requires-python = ">=3.11"
authors = [
    { name = "Temper AI Team" },
]
keywords = ["ai", "agents", "autonomous", "langgraph", "observability", "self-improvement"]
readme = "README.md"

dependencies = [
    "langchain>=1.0,<2.0",
    "langgraph>=1.0,<2.0",
    "pyyaml>=6.0,<7.0",
    "pydantic[email]>=2.0,<3.0",
    "pydantic-settings>=2.0,<3.0",
    "jinja2>=3.1.6,<4.0",
    "sqlmodel>=0.0.14,<1.0",
    "sqlalchemy>=2.0,<3.0",
    "alembic>=1.13,<2.0",
    "psycopg2-binary>=2.9,<3.0",
    "click>=8.0,<9.0",
    "rich>=13.0,<14.0",
    "httpx>=0.25,<1.0",
    "requests>=2.32.4,<3.0",
    "beautifulsoup4>=4.12,<5.0",
    "lxml>=4.9,<6.0",
    "python-dotenv>=1.0,<2.0",
    "tenacity>=8.2,<10.0",
    "cryptography>=46.0.5",
    "jsonschema>=4.0,<5.0",
]

[project.optional-dependencies]
autogen = [
    "autogen-agentchat>=0.4",
    "autogen-ext>=0.4",
]
coord = []
crewai = [
    "crewai>=0.80",
]
dashboard = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
]
dev = [
    "pytest>=7.4",
    "pytest-asyncio>=0.21",
    "pytest-cov>=4.1",
    "pytest-xdist>=3.0",
    "pytest-benchmark>=5.0",
    "hypothesis>=6.0",
    "psutil>=5.9",
    "black>=23.0",
    "ruff>=0.1",
    "mypy>=1.5",
    "bandit>=1.7",
    "radon>=6.0",
    "pip-audit>=2.6",
    "vulture>=2.10",
    "pre-commit>=3.5",
    "pytest-randomly>=3.15",
    "pytest-rerunfailures>=13.0",
    "pytest-timeout>=2.2",
    "mutmut>=2.4",
]
dspy = [
    "dspy>=2.5",
]
llm-providers = [
    "openai>=1.0",
    "anthropic>=0.8",
]
mcp = [
    "mcp>=1.26.0",
]
memory = [
    "mem0ai>=1.0.0",
]
openai_agents = [
    "openai-agents>=0.1",
]
otel = [
    "opentelemetry-api>=1.20",
    "opentelemetry-sdk>=1.20",
    "opentelemetry-exporter-otlp-proto-grpc>=1.20",
]

[project.scripts]
temper-ai = "temper_ai.interfaces.cli.main:main"

[tool.ruff]
exclude = [
    "__pycache__",
    "*.pyc",
    "*.pyo",
    "migrations",
]

[tool.ruff.format]
docstring-code-line-length = 88

[tool.ruff.lint]
select = ["E", "W", "F", "I", "C", "N", "B", "Q", "T", "UP", "PYI"]
ignore = ["E501"]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
"migrations/*" = ["E501"]

[tool.mypy]
python_version = "3.12"
warn_return_any = false
warn_unused_configs = true
ignore_missing_imports = true
